pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wavevid"
version = "0.1.0"
description = "Audio to waveform video generator with subtitles"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "click>=8.0",
    "librosa>=0.10",
    "numpy>=1.24",
    "Pillow>=10.0",
    "scipy>=1.10",
    "requests>=2.28",
]

[project.scripts]
wavevid = "wavevid.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wavevid = ["backgrounds/*.jpg", "backgrounds/*.png"]
